Provide option to re-install performance data generator on remote VMs For remote VMs, get nanotrader server base url from servlet request, if localhost then get IP address Refresh admin forms (Hyperic interface/performance data) after the request is processed

+++ spring-nanotrader-service-support/src/main/java/org/springframework/nanotrader/service/support/PerformanceRunner.java
@@ -18,16 +18,20 @@ public class PerformanceRunner implements Runnable {
 
 	private String vmPwd;
 
+	private String installOpts;
+
 	private String serverUrl;
 
 	private String numUsers;
-	
+
 	private String db;
 
-	PerformanceRunner(String numUsers, String db, String vmName, String vmUser, String vmPwd, String serverUrl) {
+	PerformanceRunner(String numUsers, String db, String vmName, String vmUser, String vmPwd, String installOpts,
+			String serverUrl) {
 		this.vmName = vmName;
 		this.vmUser = vmUser;
 		this.vmPwd = vmPwd;
+		this.installOpts = installOpts;
 		this.serverUrl = serverUrl;
 		this.numUsers = numUsers;
 		this.db = db;
@@ -36,49 +40,52 @@ public class PerformanceRunner implements Runnable {
 	@Override
 	public void run() {
 		try {
-			exec("rm -fR ~/perf-test");
-			exec("mkdir ~/perf-test");
-			exec("wget -P ~/perf-test/ "+serverUrl+"/spring-nanotrader-web/files/DataGenerator.zip");
-			exec("unzip ~/perf-test/DataGenerator.zip -d ~/perf-test/");
-			exec("rm -fr ~/performanceTest*");
-			exec("wget "+serverUrl+"/spring-nanotrader-web/files/performanceTest.sh");
-			exec("chmod 755 ~/performanceTest.sh");
+			if (installOpts.equalsIgnoreCase("true")) {
+				exec("rm -fR ~/perf-test");
+				exec("mkdir ~/perf-test");
+				exec("wget -P ~/perf-test/ " + serverUrl + "/spring-nanotrader-web/files/DataGenerator.zip");
+				exec("unzip ~/perf-test/DataGenerator.zip -d ~/perf-test/");
+				exec("rm -fr ~/performanceTest*");
+				exec("wget " + serverUrl + "/spring-nanotrader-web/files/performanceTest.sh");
+				exec("chmod 755 ~/performanceTest.sh");
+			}
 			exec("~/performanceTest.sh " + numUsers + " " + db + " " + serverUrl);
 		}
 		catch (IOException e) {
 			System.out.println(e.getMessage());
 		}
 	}
-	
-	
+
 	private class NullHostKeyVerifier implements HostKeyVerifier {
-	    @Override
-	    public boolean verify(String arg0, int arg1, PublicKey arg2) {
-	        return true;
-	    }        
+		@Override
+		public boolean verify(String arg0, int arg1, PublicKey arg2) {
+			return true;
+		}
 	}
 
-	
 	private void exec(String command) throws IOException {
 		SSHClient ssh = new SSHClient();
-	    ssh.loadKnownHosts();
-        ssh.addHostKeyVerifier(new NullHostKeyVerifier());
-	    ssh.connect(vmName);
-	    ssh.authPassword(vmUser, vmPwd);
-	    
+		ssh.loadKnownHosts();
+		ssh.addHostKeyVerifier(new NullHostKeyVerifier());
+		ssh.connect(vmName);
+		ssh.authPassword(vmUser, vmPwd);
+
 		final Session session = ssh.startSession();
-	    try {
-	      final Command cmd = session.exec(command);
-	      System.out.println(IOUtils.readFully(cmd.getInputStream()).toString());
-	      cmd.join(10, TimeUnit.SECONDS);
-	      System.out.println("\n** exit status: " + cmd.getExitStatus() + " for the cmd "+ command + " on "+ vmName);
-	      if (cmd.getExitStatus() != 0) {
-	        System.out.println("Command [" + command + "] failed with exit status :" +cmd.getExitStatus()+ " on "+vmName);
-	        //System.exit(cmd.getExitStatus());
-	      }
-	    } finally {
-	      ssh.close();
-	    }
-	    ssh.disconnect();
+		try {
+			final Command cmd = session.exec(command);
+			System.out.println(IOUtils.readFully(cmd.getInputStream()).toString());
+			cmd.join(10, TimeUnit.SECONDS);
+			System.out
+					.println("\n** exit status: " + cmd.getExitStatus() + " for the cmd " + command + " on " + vmName);
+			if (cmd.getExitStatus() != 0) {
+				System.out.println("Command [" + command + "] failed with exit status :" + cmd.getExitStatus() + " on "
+						+ vmName);
+				// System.exit(cmd.getExitStatus());
+			}
+		}
+		finally {
+			ssh.close();
+		}
+		ssh.disconnect();
 	}
 }
